created the navbar component and added it to the other pages
