+++ pages/Announcements.tsx
@@ -4,6 +4,7 @@ import {
   handleSendNotification,
   handleSendImageNotification,
 } from "../services/notificationService";
+import Navbar from "../components/navbar";
 
 const Announcements: React.FC = () => {
   const [topic, setTopic] = useState("");
@@ -38,7 +39,7 @@ const Announcements: React.FC = () => {
   };
 
   return (
-    <div className="min-h-screen bg-white flex justify-center">
+    <><Navbar></Navbar><div className="min-h-screen flex justify-center">
       <div className="w-1/2 p-8 flex flex-col">
         <p className="text-2xl font-semibold">Announcements</p>
         <label htmlFor="topic">Topic:</label>
@@ -49,8 +50,7 @@ const Announcements: React.FC = () => {
           name="topic"
           placeholder="Notification topic"
           value={topic}
-          onChange={(e) => setTopic(e.target.value)}
-        />
+          onChange={(e) => setTopic(e.target.value)} />
         <label htmlFor="title">Title:</label>
         <input
           className="bg-white shadow-md rounded px-8 pt-3 mt-3 pb-3 mb-4"
@@ -59,8 +59,7 @@ const Announcements: React.FC = () => {
           name="title"
           placeholder="Notification title"
           value={title}
-          onChange={(e) => setTitle(e.target.value)}
-        />
+          onChange={(e) => setTitle(e.target.value)} />
         <label htmlFor="body">Text:</label>
         <input
           className="bg-white shadow-md rounded px-8 pt-3 mt-3 pb-20 mb-4"
@@ -69,8 +68,7 @@ const Announcements: React.FC = () => {
           name="body"
           placeholder="Enter text here..."
           value={body}
-          onChange={(e) => setBody(e.target.value)}
-        />
+          onChange={(e) => setBody(e.target.value)} />
       </div>
       <div className="w-1/2 p-8 flex flex-col">
         <label htmlFor="imageUrl">Image URL:</label>
@@ -82,8 +80,7 @@ const Announcements: React.FC = () => {
           placeholder="Image URL"
           autoComplete="off"
           value={imageUrl}
-          onChange={(e) => setImageUrl(e.target.value)}
-        />
+          onChange={(e) => setImageUrl(e.target.value)} />
         <button
           type="button" // Use type="button" to prevent form submission
           className="px-4 py-2 font-bold text-white bg-blue-500 rounded hover-bg-blue-700 focus:outline-none focus:shadow-outline"
@@ -92,7 +89,7 @@ const Announcements: React.FC = () => {
           Send Notification
         </button>
       </div>
-    </div>
+    </div></>
   );
 };
