+++ pages/_app.tsx
@@ -1,9 +1,12 @@
 import '../app/globals.css';
 import { AppProps } from 'next/app';
+import { Routes, Route } from 'react-router-dom'
+import Navbar from '../components/navbar';
 
 function MyApp({ Component, pageProps }: AppProps) {
 
     return <Component {...pageProps} />;
+    
 }
 
 export default MyApp;
